+++ sdk/resourcemanager/storagecache/armstoragecache/models.go
@@ -388,6 +388,117 @@ type AscOperationProperties struct {
 	Output map[string]any
 }
 
+// AutoExportJob - An auto export job instance. Follows Azure Resource Manager standards: https://github.com/Azure/azure-resource-manager-rpc/blob/master/v1.0/resource-api-reference.md
+type AutoExportJob struct {
+	// REQUIRED; The geo-location where the resource lives
+	Location *string
+
+	// Properties of the auto export job.
+	Properties *AutoExportJobProperties
+
+	// Resource tags.
+	Tags map[string]*string
+
+	// READ-ONLY; Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}
+	ID *string
+
+	// READ-ONLY; The name of the resource
+	Name *string
+
+	// READ-ONLY; Azure Resource Manager metadata containing createdBy and modifiedBy information.
+	SystemData *SystemData
+
+	// READ-ONLY; The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts"
+	Type *string
+}
+
+// AutoExportJobProperties - Properties of the auto export job.
+type AutoExportJobProperties struct {
+	// The administrative status of the auto export job. Possible values: 'Enable', 'Disable'. Passing in a value of 'Disable'
+	// will disable the current active auto export job. By default it is set to
+	// 'Enable'.
+	AdminStatus *AutoExportJobAdminStatus
+
+	// An array of blob paths/prefixes that get auto exported to the cluster namespace. It has '/' as the default value. Number
+	// of maximum allowed paths for now is 1.
+	AutoExportPrefixes []*string
+
+	// READ-ONLY; ARM provisioning state.
+	ProvisioningState *AutoExportJobProvisioningStateType
+
+	// READ-ONLY; The status of the auto export
+	Status *AutoExportJobPropertiesStatus
+}
+
+// AutoExportJobPropertiesStatus - The status of the auto export
+type AutoExportJobPropertiesStatus struct {
+	// The operational state of auto export. InProgress indicates the export is running. Disabling indicates the user has requested
+	// to disable the export but the disabling is still in progress. Disabled
+	// indicates auto export has been disabled. DisableFailed indicates the disabling has failed. Failed means the export was
+	// unable to continue, due to a fatal error.
+	State *AutoExportStatusType
+
+	// READ-ONLY; Files discovered for export in current iteration. It may increase while more export items are found.
+	CurrentIterationFilesDiscovered *int64
+
+	// READ-ONLY; Number of files that have been exported in current iteration.
+	CurrentIterationFilesExported *int64
+
+	// READ-ONLY; Number of files failed to export in current iteration.
+	CurrentIterationFilesFailed *int64
+
+	// READ-ONLY; Data (in MiB) discovered for export in current iteration. It may increase while more export items are found.
+	CurrentIterationMiBDiscovered *int64
+
+	// READ-ONLY; Data (in MiB) that have been exported in current iteration.
+	CurrentIterationMiBExported *int64
+
+	// READ-ONLY; Number of iterations completed since the start of the export.
+	ExportIterationCount *int32
+
+	// READ-ONLY; The time (in UTC) of the last completed auto export job.
+	LastCompletionTimeUTC *time.Time
+
+	// READ-ONLY; The time (in UTC) the latest auto export job started.
+	LastStartedTimeUTC *time.Time
+
+	// READ-ONLY; Time (in UTC) of the last successfully completed export iteration. Look at logging container for details.
+	LastSuccessfulIterationCompletionTimeUTC *time.Time
+
+	// READ-ONLY; Server-defined status code for auto export job.
+	StatusCode *string
+
+	// READ-ONLY; Server-defined status message for auto export job.
+	StatusMessage *string
+
+	// READ-ONLY; Total files exported since the start of the export. This is accumulative, some files may be counted repeatedly.
+	TotalFilesExported *int64
+
+	// READ-ONLY; Total files failed to be export since the last successfully completed iteration. This is accumulative, some
+	// files may be counted repeatedly.
+	TotalFilesFailed *int64
+
+	// READ-ONLY; Total data (in MiB) exported since the start of the export. This is accumulative, some files may be counted
+	// repeatedly.
+	TotalMiBExported *int64
+}
+
+// AutoExportJobUpdate - An auto export job update instance.
+type AutoExportJobUpdate struct {
+	// Resource tags.
+	Tags map[string]*string
+}
+
+// AutoExportJobsListResult - Result of the request to list auto export jobs. It contains a list of auto export jobs and a
+// URL link to get the next set of results.
+type AutoExportJobsListResult struct {
+	// URL to get the next set of auto export job list results, if there are any.
+	NextLink *string
+
+	// List of auto export jobs.
+	Value []*AutoExportJob
+}
+
 // BlobNfsTarget - Properties pertaining to the BlobNfsTarget.
 type BlobNfsTarget struct {
 	// Resource ID of the storage container.
@@ -726,6 +837,10 @@ type ImportJob struct {
 
 // ImportJobProperties - Properties of the import job.
 type ImportJobProperties struct {
+	// The administrative status of the import job. Possible values: 'Enable', 'Disable'. Passing in a value of 'Disable' will
+	// cancel the current active import job. By default it is set to 'Enable'.
+	AdminStatus *ImportJobAdminStatus
+
 	// How the import job will handle conflicts. For example, if the import job is trying to bring in a directory, but a file
 	// is at that path, how it handles it. Fail indicates that the import job should
 	// stop immediately and not do anything with the conflict. Skip indicates that it should pass over the conflict. OverwriteIfDirty
@@ -757,17 +872,35 @@ type ImportJobPropertiesStatus struct {
 	// READ-ONLY; A recent and frequently updated rate of blobs walked per second.
 	BlobsWalkedPerSecond *int64
 
-	// READ-ONLY; The time of the last completed archive operation
+	// READ-ONLY; Number of new or modified directories that have been imported into the filesystem.
+	ImportedDirectories *int64
+
+	// READ-ONLY; Number of new or modified files that have been imported into the filesystem.
+	ImportedFiles *int64
+
+	// READ-ONLY; Number of newly added symbolic links into the filesystem.
+	ImportedSymlinks *int64
+
+	// READ-ONLY; The time (in UTC) of the last completed import job.
 	LastCompletionTime *time.Time
 
-	// READ-ONLY; The time the latest archive operation started
+	// READ-ONLY; The time (in UTC) the latest import job started.
 	LastStartedTime *time.Time
 
-	// READ-ONLY; The state of the import job. InProgress indicates the import is still running. Canceled indicates it has been
-	// canceled by the user. Completed indicates import finished, successfully importing all
-	// discovered blobs into the Lustre namespace. CompletedPartial indicates the import finished but some blobs either were found
-	// to be conflicting and could not be imported or other errors were
-	// encountered. Failed means the import was unable to complete due to a fatal error.
+	// READ-ONLY; Number of directories that already exist in the filesystem and have not been modified.
+	PreexistingDirectories *int64
+
+	// READ-ONLY; Number of files that already exist in the filesystem and have not been modified.
+	PreexistingFiles *int64
+
+	// READ-ONLY; Number of symbolic links that already exist in the filesystem and have not been modified.
+	PreexistingSymlinks *int64
+
+	// READ-ONLY; The operational state of the import job. InProgress indicates the import is still running. Canceled indicates
+	// it has been canceled by the user. Completed indicates import finished, successfully
+	// importing all discovered blobs into the Lustre namespace. CompletedPartial indicates the import finished but some blobs
+	// either were found to be conflicting and could not be imported or other errors
+	// were encountered. Failed means the import was unable to complete due to a fatal error.
 	State *ImportStatusType
 
 	// READ-ONLY; The status message of the import job.