+++ sdk/resourcemanager/storagecache/armstoragecache/models_serde.go
@@ -1012,6 +1012,237 @@ func (a *AscOperationProperties) UnmarshalJSON(data []byte) error {
 	return nil
 }
 
+// MarshalJSON implements the json.Marshaller interface for type AutoExportJob.
+func (a AutoExportJob) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "id", a.ID)
+	populate(objectMap, "location", a.Location)
+	populate(objectMap, "name", a.Name)
+	populate(objectMap, "properties", a.Properties)
+	populate(objectMap, "systemData", a.SystemData)
+	populate(objectMap, "tags", a.Tags)
+	populate(objectMap, "type", a.Type)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AutoExportJob.
+func (a *AutoExportJob) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "id":
+			err = unpopulate(val, "ID", &a.ID)
+			delete(rawMsg, key)
+		case "location":
+			err = unpopulate(val, "Location", &a.Location)
+			delete(rawMsg, key)
+		case "name":
+			err = unpopulate(val, "Name", &a.Name)
+			delete(rawMsg, key)
+		case "properties":
+			err = unpopulate(val, "Properties", &a.Properties)
+			delete(rawMsg, key)
+		case "systemData":
+			err = unpopulate(val, "SystemData", &a.SystemData)
+			delete(rawMsg, key)
+		case "tags":
+			err = unpopulate(val, "Tags", &a.Tags)
+			delete(rawMsg, key)
+		case "type":
+			err = unpopulate(val, "Type", &a.Type)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type AutoExportJobProperties.
+func (a AutoExportJobProperties) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "adminStatus", a.AdminStatus)
+	populate(objectMap, "autoExportPrefixes", a.AutoExportPrefixes)
+	populate(objectMap, "provisioningState", a.ProvisioningState)
+	populate(objectMap, "status", a.Status)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AutoExportJobProperties.
+func (a *AutoExportJobProperties) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "adminStatus":
+			err = unpopulate(val, "AdminStatus", &a.AdminStatus)
+			delete(rawMsg, key)
+		case "autoExportPrefixes":
+			err = unpopulate(val, "AutoExportPrefixes", &a.AutoExportPrefixes)
+			delete(rawMsg, key)
+		case "provisioningState":
+			err = unpopulate(val, "ProvisioningState", &a.ProvisioningState)
+			delete(rawMsg, key)
+		case "status":
+			err = unpopulate(val, "Status", &a.Status)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type AutoExportJobPropertiesStatus.
+func (a AutoExportJobPropertiesStatus) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "currentIterationFilesDiscovered", a.CurrentIterationFilesDiscovered)
+	populate(objectMap, "currentIterationFilesExported", a.CurrentIterationFilesExported)
+	populate(objectMap, "currentIterationFilesFailed", a.CurrentIterationFilesFailed)
+	populate(objectMap, "currentIterationMiBDiscovered", a.CurrentIterationMiBDiscovered)
+	populate(objectMap, "currentIterationMiBExported", a.CurrentIterationMiBExported)
+	populate(objectMap, "exportIterationCount", a.ExportIterationCount)
+	populateDateTimeRFC3339(objectMap, "lastCompletionTimeUTC", a.LastCompletionTimeUTC)
+	populateDateTimeRFC3339(objectMap, "lastStartedTimeUTC", a.LastStartedTimeUTC)
+	populateDateTimeRFC3339(objectMap, "lastSuccessfulIterationCompletionTimeUTC", a.LastSuccessfulIterationCompletionTimeUTC)
+	populate(objectMap, "state", a.State)
+	populate(objectMap, "statusCode", a.StatusCode)
+	populate(objectMap, "statusMessage", a.StatusMessage)
+	populate(objectMap, "totalFilesExported", a.TotalFilesExported)
+	populate(objectMap, "totalFilesFailed", a.TotalFilesFailed)
+	populate(objectMap, "totalMiBExported", a.TotalMiBExported)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AutoExportJobPropertiesStatus.
+func (a *AutoExportJobPropertiesStatus) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "currentIterationFilesDiscovered":
+			err = unpopulate(val, "CurrentIterationFilesDiscovered", &a.CurrentIterationFilesDiscovered)
+			delete(rawMsg, key)
+		case "currentIterationFilesExported":
+			err = unpopulate(val, "CurrentIterationFilesExported", &a.CurrentIterationFilesExported)
+			delete(rawMsg, key)
+		case "currentIterationFilesFailed":
+			err = unpopulate(val, "CurrentIterationFilesFailed", &a.CurrentIterationFilesFailed)
+			delete(rawMsg, key)
+		case "currentIterationMiBDiscovered":
+			err = unpopulate(val, "CurrentIterationMiBDiscovered", &a.CurrentIterationMiBDiscovered)
+			delete(rawMsg, key)
+		case "currentIterationMiBExported":
+			err = unpopulate(val, "CurrentIterationMiBExported", &a.CurrentIterationMiBExported)
+			delete(rawMsg, key)
+		case "exportIterationCount":
+			err = unpopulate(val, "ExportIterationCount", &a.ExportIterationCount)
+			delete(rawMsg, key)
+		case "lastCompletionTimeUTC":
+			err = unpopulateDateTimeRFC3339(val, "LastCompletionTimeUTC", &a.LastCompletionTimeUTC)
+			delete(rawMsg, key)
+		case "lastStartedTimeUTC":
+			err = unpopulateDateTimeRFC3339(val, "LastStartedTimeUTC", &a.LastStartedTimeUTC)
+			delete(rawMsg, key)
+		case "lastSuccessfulIterationCompletionTimeUTC":
+			err = unpopulateDateTimeRFC3339(val, "LastSuccessfulIterationCompletionTimeUTC", &a.LastSuccessfulIterationCompletionTimeUTC)
+			delete(rawMsg, key)
+		case "state":
+			err = unpopulate(val, "State", &a.State)
+			delete(rawMsg, key)
+		case "statusCode":
+			err = unpopulate(val, "StatusCode", &a.StatusCode)
+			delete(rawMsg, key)
+		case "statusMessage":
+			err = unpopulate(val, "StatusMessage", &a.StatusMessage)
+			delete(rawMsg, key)
+		case "totalFilesExported":
+			err = unpopulate(val, "TotalFilesExported", &a.TotalFilesExported)
+			delete(rawMsg, key)
+		case "totalFilesFailed":
+			err = unpopulate(val, "TotalFilesFailed", &a.TotalFilesFailed)
+			delete(rawMsg, key)
+		case "totalMiBExported":
+			err = unpopulate(val, "TotalMiBExported", &a.TotalMiBExported)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type AutoExportJobUpdate.
+func (a AutoExportJobUpdate) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "tags", a.Tags)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AutoExportJobUpdate.
+func (a *AutoExportJobUpdate) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "tags":
+			err = unpopulate(val, "Tags", &a.Tags)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
+// MarshalJSON implements the json.Marshaller interface for type AutoExportJobsListResult.
+func (a AutoExportJobsListResult) MarshalJSON() ([]byte, error) {
+	objectMap := make(map[string]any)
+	populate(objectMap, "nextLink", a.NextLink)
+	populate(objectMap, "value", a.Value)
+	return json.Marshal(objectMap)
+}
+
+// UnmarshalJSON implements the json.Unmarshaller interface for type AutoExportJobsListResult.
+func (a *AutoExportJobsListResult) UnmarshalJSON(data []byte) error {
+	var rawMsg map[string]json.RawMessage
+	if err := json.Unmarshal(data, &rawMsg); err != nil {
+		return fmt.Errorf("unmarshalling type %T: %v", a, err)
+	}
+	for key, val := range rawMsg {
+		var err error
+		switch key {
+		case "nextLink":
+			err = unpopulate(val, "NextLink", &a.NextLink)
+			delete(rawMsg, key)
+		case "value":
+			err = unpopulate(val, "Value", &a.Value)
+			delete(rawMsg, key)
+		}
+		if err != nil {
+			return fmt.Errorf("unmarshalling type %T: %v", a, err)
+		}
+	}
+	return nil
+}
+
 // MarshalJSON implements the json.Marshaller interface for type BlobNfsTarget.
 func (b BlobNfsTarget) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
@@ -1823,6 +2054,7 @@ func (i *ImportJob) UnmarshalJSON(data []byte) error {
 // MarshalJSON implements the json.Marshaller interface for type ImportJobProperties.
 func (i ImportJobProperties) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
+	populate(objectMap, "adminStatus", i.AdminStatus)
 	populate(objectMap, "conflictResolutionMode", i.ConflictResolutionMode)
 	populate(objectMap, "importPrefixes", i.ImportPrefixes)
 	populate(objectMap, "maximumErrors", i.MaximumErrors)
@@ -1840,6 +2072,9 @@ func (i *ImportJobProperties) UnmarshalJSON(data []byte) error {
 	for key, val := range rawMsg {
 		var err error
 		switch key {
+		case "adminStatus":
+			err = unpopulate(val, "AdminStatus", &i.AdminStatus)
+			delete(rawMsg, key)
 		case "conflictResolutionMode":
 			err = unpopulate(val, "ConflictResolutionMode", &i.ConflictResolutionMode)
 			delete(rawMsg, key)
@@ -1868,8 +2103,14 @@ func (i ImportJobPropertiesStatus) MarshalJSON() ([]byte, error) {
 	objectMap := make(map[string]any)
 	populate(objectMap, "blobsImportedPerSecond", i.BlobsImportedPerSecond)
 	populate(objectMap, "blobsWalkedPerSecond", i.BlobsWalkedPerSecond)
+	populate(objectMap, "importedDirectories", i.ImportedDirectories)
+	populate(objectMap, "importedFiles", i.ImportedFiles)
+	populate(objectMap, "importedSymlinks", i.ImportedSymlinks)
 	populateDateTimeRFC3339(objectMap, "lastCompletionTime", i.LastCompletionTime)
 	populateDateTimeRFC3339(objectMap, "lastStartedTime", i.LastStartedTime)
+	populate(objectMap, "preexistingDirectories", i.PreexistingDirectories)
+	populate(objectMap, "preexistingFiles", i.PreexistingFiles)
+	populate(objectMap, "preexistingSymlinks", i.PreexistingSymlinks)
 	populate(objectMap, "state", i.State)
 	populate(objectMap, "statusMessage", i.StatusMessage)
 	populate(objectMap, "totalBlobsImported", i.TotalBlobsImported)
@@ -1894,11 +2135,29 @@ func (i *ImportJobPropertiesStatus) UnmarshalJSON(data []byte) error {
 		case "blobsWalkedPerSecond":
 			err = unpopulate(val, "BlobsWalkedPerSecond", &i.BlobsWalkedPerSecond)
 			delete(rawMsg, key)
+		case "importedDirectories":
+			err = unpopulate(val, "ImportedDirectories", &i.ImportedDirectories)
+			delete(rawMsg, key)
+		case "importedFiles":
+			err = unpopulate(val, "ImportedFiles", &i.ImportedFiles)
+			delete(rawMsg, key)
+		case "importedSymlinks":
+			err = unpopulate(val, "ImportedSymlinks", &i.ImportedSymlinks)
+			delete(rawMsg, key)
 		case "lastCompletionTime":
 			err = unpopulateDateTimeRFC3339(val, "LastCompletionTime", &i.LastCompletionTime)
 			delete(rawMsg, key)
 		case "lastStartedTime":
 			err = unpopulateDateTimeRFC3339(val, "LastStartedTime", &i.LastStartedTime)
+			delete(rawMsg, key)
+		case "preexistingDirectories":
+			err = unpopulate(val, "PreexistingDirectories", &i.PreexistingDirectories)
+			delete(rawMsg, key)
+		case "preexistingFiles":
+			err = unpopulate(val, "PreexistingFiles", &i.PreexistingFiles)
+			delete(rawMsg, key)
+		case "preexistingSymlinks":
+			err = unpopulate(val, "PreexistingSymlinks", &i.PreexistingSymlinks)
 			delete(rawMsg, key)
 		case "state":
 			err = unpopulate(val, "State", &i.State)