+++ sdk/resourcemanager/storagecache/armstoragecache/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Release History
 
+## 4.1.0 (2024-06-24)
+### Features Added
+
+- New enum type `AutoExportJobAdminStatus` with values `AutoExportJobAdminStatusDisable`, `AutoExportJobAdminStatusEnable`
+- New enum type `AutoExportJobProvisioningStateType` with values `AutoExportJobProvisioningStateTypeCanceled`, `AutoExportJobProvisioningStateTypeCreating`, `AutoExportJobProvisioningStateTypeDeleting`, `AutoExportJobProvisioningStateTypeFailed`, `AutoExportJobProvisioningStateTypeSucceeded`, `AutoExportJobProvisioningStateTypeUpdating`
+- New enum type `AutoExportStatusType` with values `AutoExportStatusTypeDisableFailed`, `AutoExportStatusTypeDisabled`, `AutoExportStatusTypeDisabling`, `AutoExportStatusTypeFailed`, `AutoExportStatusTypeInProgress`
+- New enum type `ImportJobAdminStatus` with values `ImportJobAdminStatusDisable`, `ImportJobAdminStatusEnable`
+- New function `NewAutoExportJobsClient(string, azcore.TokenCredential, *arm.ClientOptions) (*AutoExportJobsClient, error)`
+- New function `*AutoExportJobsClient.BeginCreateOrUpdate(context.Context, string, string, string, AutoExportJob, *AutoExportJobsClientBeginCreateOrUpdateOptions) (*runtime.Poller[AutoExportJobsClientCreateOrUpdateResponse], error)`
+- New function `*AutoExportJobsClient.BeginDelete(context.Context, string, string, string, *AutoExportJobsClientBeginDeleteOptions) (*runtime.Poller[AutoExportJobsClientDeleteResponse], error)`
+- New function `*AutoExportJobsClient.Get(context.Context, string, string, string, *AutoExportJobsClientGetOptions) (AutoExportJobsClientGetResponse, error)`
+- New function `*AutoExportJobsClient.NewListByAmlFilesystemPager(string, string, *AutoExportJobsClientListByAmlFilesystemOptions) *runtime.Pager[AutoExportJobsClientListByAmlFilesystemResponse]`
+- New function `*AutoExportJobsClient.BeginUpdate(context.Context, string, string, string, AutoExportJobUpdate, *AutoExportJobsClientBeginUpdateOptions) (*runtime.Poller[AutoExportJobsClientUpdateResponse], error)`
+- New function `*ClientFactory.NewAutoExportJobsClient() *AutoExportJobsClient`
+- New struct `AutoExportJob`
+- New struct `AutoExportJobProperties`
+- New struct `AutoExportJobPropertiesStatus`
+- New struct `AutoExportJobUpdate`
+- New struct `AutoExportJobsListResult`
+- New field `AdminStatus` in struct `ImportJobProperties`
+- New field `ImportedDirectories`, `ImportedFiles`, `ImportedSymlinks`, `PreexistingDirectories`, `PreexistingFiles`, `PreexistingSymlinks` in struct `ImportJobPropertiesStatus`
+
+
 ## 4.0.0 (2024-05-24)
 ### Breaking Changes